Refactor/backend

+++ backend/app/Http/Controllers/LayoutController.php
@@ -8,6 +8,7 @@ use App\Http\Requests\StoreLayoutRequest;
 use App\Http\Requests\UpdateLayoutRequest;
 use App\Http\Resources\LayoutResource;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Log;
 
 /**
  * レイアウトに関する操作を管理するコントローラークラスです。
@@ -38,7 +39,7 @@ class LayoutController extends Controller
      */
     public function store(StoreLayoutRequest $request)
     {
-        $data = $request->only(['text', 'itemIds']);
+        $data = $request->only(['text', 'items']);
         $this->layoutService->createLayout($data, Auth::id());
         return response()->json(['message' => 'レイアウト登録が完了しました。'], 201);
     }
@@ -61,7 +62,8 @@ class LayoutController extends Controller
      */
     public function update(UpdateLayoutRequest $request, Layout $layout)
     {
-        $data = $request->only(['text', 'itemIds']);
+        $this->authorize('update', $layout);
+        $data = $request->only(['text', 'items']);
         $this->layoutService->updateLayout($layout, $data);
         return response()->json(['message' => 'レイアウト更新が完了しました。'], 200);
     }
@@ -72,6 +74,7 @@ class LayoutController extends Controller
      */
     public function destroy(Layout $layout)
     {
+        $this->authorize('delete', $layout);
         $this->layoutService->removeLayout($layout);
         return response(null, 204);
     }
